requirements.txt
streamlit
torch
torchvision
diffusers
transformers
pillow
numpy
scikit-image

